pipeline/pyproject.toml
[project]
name = "resume-pipeline"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "langgraph>=0.2.73",
    "langchain-anthropic>=0.3.3",
    "langchain-openai>=0.3.0",
    "instructor>=1.7.2",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    "guardrails-ai>=0.6.5",
    "llmlingua>=0.2.2",
    "supabase>=2.10.0",
    "upstash-redis>=1.3.0",
    "azure-servicebus>=7.12.0",
    "azure-storage-blob>=12.22.0",
    "langfuse>=2.60.0",
    "python-multipart>=0.0.20",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "openai>=1.59.0",
    "anthropic>=0.43.0",
    "pypdf>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
